- LAP_ImplBase as an abstract class.
